ptc/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "klinklang"
version = "0.0.1"
description = "The public components for klinklang"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Christopher Müller", email = "christphermueller2@outlook.de"},
    {name = "James Wilson", email = "contact@james.baby"},
]
dependencies = ["pymongo==4.6.3","PyYaml==6.0.1","pydantic==1.10.14",]

[project.urls]
Homepage = "https://github.com/SpielerNogard/klinklang-public"

[tool.setuptools.packages.find]
include = ["klinklang*"]
exclude=["*.tests", "*.tests.*", "tests.*", "tests"]

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py38"

[tool.ruff.lint]
# for more rules: https://docs.astral.sh/ruff/rules/
select = [
    "A",  # prevent using keywords that clobber python builtins
    "E",  # pycodestyle
    "F",  # pyflakes
    "UP",  # alert you when better syntax is available in your python version
    "I",  # isort
    "N",   # PEP8 naming
    "W"   # Warnings
]
ignore = ["A003","N805","A002"]

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"
# Like Black, indent with spaces, rather than tabs.
indent-style = "space"
# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false
# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"
# Enable reformatting of code snippets in docstrings.
docstring-code-format = true

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "semver"
version_provider = "pep621"
update_changelog_on_bump = true
bump_message = "bump: $current_version → $new_version [skip-ci]"
version_files = [
    "klinklang/__init__.py",
]